Create Strategic Ecosystems controller
