+++ src/pages/search/dashboard/EcosystemsController.ts
@@ -114,44 +114,6 @@ export class EcosystemsController {
     throw Error("Polygon and area undefined");
   }
 
-  /**
-   * Get the raster layers required for a Special Ecosystem type
-   *  @param {string} seType Special Ecosystem type
-   *
-   * @returns { Promise<Array<RasterLayer>> } layers for the Special Ecosystem type
-   */
-
-  /**
-   * Get the coverage values for the current area
-   *
-   * @returns { Promise<SmallStackedBarData[]>}
-   */
-  async getStrategicEcosystemsValues(): Promise<SEData[]> {
-    const keys = ["paramo", "tropicalDryForest", "wetland"] as const;
-
-    const requests = keys.map((key) =>
-      SearchAPI.requestMetricsValues<typeof key>(key, this.areaId),
-    );
-
-    const responses = await Promise.all(requests);
-
-    const result: SEData[] = responses.map((res, index) => {
-      const { id, ...values } = res;
-
-      const totalArea = Object.values(values).reduce(
-        (acc, value) => acc + value,
-        0,
-      );
-
-      return {
-        type: keys[index],
-        area: Number(totalArea),
-        values,
-      };
-    });
-
-    return result;
-  }
   // TODO: Refactor to use SearchAPI when available
 
   /*
